fix: 개인정보 관리 페이지 data fetch 로직 하위 컴포넌트로 분리
